refactor: 최근 조회 숙소와 위시리스트 목록 반환 dto에서 숙소 전체 이미지를 반환하는 것에서 썸네일만 반환하도록 변경

+++ src/main/java/kr/kro/airbob/domain/review/dto/ReviewResponse.java
@@ -50,7 +50,7 @@ public class ReviewResponse {
 				return new ReviewSummary(0, BigDecimal.ZERO);
 			}
 			return new ReviewSummary(
-				summary.getTotalReviewCount().intValue(),
+				summary.getTotalReviewCount(),
 				summary.getAverageRating()
 			);
 		}
